+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/ui/Util.kt
@@ -135,15 +135,6 @@ fun Panel.ccv2ServiceReplicasRow(service: CCv2ServiceDto) {
     }
 }
 
-fun Panel.ccv2ServiceModifiedByRow(service: CCv2ServiceDto) {
-    row {
-        icon(HybrisIcons.CCv2.Service.MODIFIED_BY)
-            .gap(RightGap.SMALL)
-        label(service.modifiedBy)
-            .comment("Modified by")
-    }
-}
-
 fun Panel.ccv2ServiceModifiedTimeRow(service: CCv2ServiceDto) {
     row {
         date("Modified time", service.modifiedTime)