+++ modules/ccv2/ui/src/sap/commerce/toolset/ccv2/ui/view/CCv2EnvironmentDetailsView.kt
@@ -266,7 +266,9 @@ class CCv2EnvironmentDetailsView(
                     .gap(RightGap.COLUMNS)
 
                 panel {
-                    ccv2ServiceModifiedByRow(service)
+                    row {
+                        sUser(project, service.modifiedBy, HybrisIcons.CCv2.Service.MODIFIED_BY, "Modified by")
+                    }
                 }
                     .gap(RightGap.COLUMNS)